+++ index.html
@@ -5,37 +5,25 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <script defer src="./script.js"></script>
+  <!-- <script type = 'module' defer src="./main.js"></script> -->
   <link rel="stylesheet" href="./style.css">
   <title>PuppyBowl</title>
 </head>
 
 <body>
-  <header>
-    <h1>Welcome to the 2023 Puppy Bowl!!!</h1>
-    <p class="welcome">Welcome to the Puppy Bowl 2023, the cutest sporting event of the year! Get ready for a paw-some
-      display of puppy athleticism, where fur and fluff collide in an epic battle for the title of the Most Adorable
-      Athlete.But don't be fooled by their innocent looks – these pups are fierce competitors. They'll go head-to-head,
-      barking
-      their way to victory, and proving that size doesn't matter when it comes to determination and puppy power.
-      The halftime show promises to be an absolute treat, featuring the world-famous Puppy Marching Band and their
-      adorable rendition of "Who Let the Dogs Out." And let's not forget the Puppy Bowl cheerleaders, bringing their
-      infectious energy and tail-wagging enthusiasm to the sidelines.
-      So grab your snacks, snuggle up with your favorite four-legged friend, and prepare to be overwhelmed by an
-      avalanche of cuteness. Puppy Bowl 2023 is an event you won't want to miss, where the stakes are high, the fur is
-      fluffy, and the "awwws" are in abundance. Get ready for a barking good time!</p>
-  </header>
-  <h2>Add New Puppy Bowl Player</h2>
+ 
+  
   <div id="new-player-form"></div>
   <!-- Participants will be dynamically rendered here -->
   <h2>Meet This Years Puppers</h2>
   <!-- container for all players -->
-  <div id="all-players-container"></div>
+  <div id="all-players-container";></div>
   <!-- form to add new players -->
 
 
   <!-- import JavaScript -->
-  <script src="./script.js"></script>
+  
+  <script src="./main.js" type="module"></script>
 </body>
 
 </html>
